pyproject.toml
[project]
name = "aggregator-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "firecrawl>=2.16.5",
    "httpx>=0.28.1",
    "openai>=1.99.2",
    "openpyxl>=3.1.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
